cpu: inline Prospero sparse page tables

+++ src/dynarmic/src/dynarmic/backend/x64/emit_x64_data_processing.cpp
@@ -930,7 +930,8 @@ static void EmitAdd(BlockOfCode& code, EmitContext& ctx, IR::Inst* inst, size_t 
     if (!carry_inst && !overflow_inst && !nzcv_inst && carry_in.IsImmediate() && !carry_in.GetImmediateU1()) {
         if (args[1].IsImmediate() && args[1].FitsInImmediateS32()) {
             Xbyak::Reg const result = ctx.reg_alloc.UseScratchGpr(code, args[0]).changeBit(bitsize);
-            code.lea(result, code.ptr[result + args[1].GetImmediateS32()]);
+            code.lea(result,
+                     code.ptr[result + static_cast<std::int32_t>(args[1].GetImmediateS32())]);
             ctx.reg_alloc.DefineValue(code, inst, result);
         } else {
             Xbyak::Reg const result = ctx.reg_alloc.UseScratchGpr(code, args[0]).changeBit(bitsize);
@@ -956,7 +957,7 @@ static void EmitAdd(BlockOfCode& code, EmitContext& ctx, IR::Inst* inst, size_t 
                     code.stc();
                     code.adc(result, op_arg);
                 } else {
-                    code.lea(result, code.ptr[result + op_arg + 1]);
+                    code.lea(result, code.ptr[result + static_cast<std::int32_t>(op_arg + 1)]);
                 }
             } else {
                 code.add(result, op_arg);
